テストスクリプトを追加する

+++ mongodb/stop.sh
@@ -12,7 +12,7 @@ dir=$currentDir/versions/$optVersion
 
 exitIfNotExistDir $dir/datadir/$optName
 exitIfNotRunningPort $optPort
-kill `cat $dir/datadir/$optName/mongodb.pid` && rm -f $dir/datadir/$optName/mongodb.pid
-cp $dir/datadir/$optName/mongodb.port $dir/datadir/$optName/mongodb.port.last
-rm -f $dir/datadir/$optName/mongodb.port
+[ -f "$dir/datadir/$optName/mongodb.pid" ] && kill $(cat $dir/datadir/$optName/mongodb.pid) && rm -f $dir/datadir/$optName/mongodb.pid
+[ -f "$dir/datadir/$optName/mongodb.port" ] && cp $dir/datadir/$optName/mongodb.port $dir/datadir/$optName/mongodb.port.last
+[ -f "$dir/datadir/$optName/mongodb.port" ] && rm -f $dir/datadir/$optName/mongodb.port
 echo MongoDB Successfully stopped. $optName $optVersion $optPort
